requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.3
djangorestframework==3.15.1
drf-yasg==1.21.7
filelock==3.13.3
inflection==0.5.1
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
psycopg2==2.9.9
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
